split DocumentEditor/divider into shared/ui
