+++ packages/fields-document/src/DocumentEditor/block-markdown-shortcuts.ts
@@ -1,7 +1,7 @@
 import { Element, Editor, Transforms, Range } from 'slate'
 import { type DocumentFeatures } from '../views'
 import { type ComponentBlock } from './component-blocks/api-shared'
-import { insertDivider } from './divider'
+import { insertDivider } from './divider-shared'
 import { type DocumentFeaturesForNormalization } from './document-features-normalization'
 import { getAncestorComponentChildFieldDocumentFeatures } from './toolbar-state'
