+++ packages/fields-document/src/DocumentEditor/index.tsx
@@ -35,7 +35,7 @@ import { ComponentBlockContext, withComponentBlocks } from './component-blocks'
 import { getPlaceholderTextForPropPath } from './component-blocks/utils'
 import { withBlockquote } from './blockquote-shared'
 import { type Relationships, withRelationship } from './relationship'
-import { withDivider } from './divider'
+import { withDivider } from './divider-shared'
 import { withCodeBlock } from './code-block-shared'
 import { withMarks } from './marks'
 import { renderLeaf } from './leaf'
